Report helpful error message on invalid Scala version.
Previously, the `nativeImage` task failed with the follow error when
using Scala 2.12.10

```
[error] Error: substitution target for scala.meta.internal.svm_subs.Target_scala_runtime_Statics is not loaded. Use field `onlyWith` in the `TargetClass` annotation to make substitution only active when needed.
```

Now, the error message is
```
outdated Scala version. To fix this problem, add `scalaVersion := "2.12.12"` (or any version newer than 2.12.12) to build.sbt
```

A better fix would be to use the `onlyWith` field as recommended in the
native-image error message. However, I failed to get this working.