Use improved native-image substitutions
Fixes #12.

Previously, native-image generation failed if the user had an older
version of Scala like 2.12.10. Now, native-image generation works with
any Scala version thanks to improvements in the upstream native-image
substitutions. The substitutions are now conditionally enabled based on
the presence of `releaseFence()` methods. The substitions now have no
effect if they are not needed.